feat: create getGifs helper
