+++ src/components/AddCategory.js
@@ -11,7 +11,7 @@ export const AddCategory = ({ setCategories }) => {
   const handleSubmit = (e) => {
     e.preventDefault();
     if (inputValue.trim().length > 2) {
-      setCategories(cats => [...cats, inputValue]);
+      setCategories(cats => [inputValue, ...cats]);
       setInputValue('');
     }
   }
